test: revert test_unit_test.py to original doc example state and maintain separate test files
